@@ -51,7 +51,7 @@ class UserController extends Controller {
     }
 
     private function createCreateForm(User $entity) {
-        $form = $this->createForm(new UserType(), $entity, array('action' => $this->generateUrl('user_create'),
+        $form = $this->createForm(UserType::class, $entity, array('action' => $this->generateUrl('user_create'),
             'method' => 'POST'
         ));
         return $form;
@@ -113,7 +113,7 @@ class UserController extends Controller {
 
     private function createEditForm(User $entity) {
         $form = $this->createForm(
-                new UserType(), $entity, array('action' => $this->generateUrl('user_update', array('id' => $entity->getId())), 'method' => 'PUT')
+                UserType::class, $entity, array('action' => $this->generateUrl('user_update', array('id' => $entity->getId())), 'method' => 'PUT')
         );
 
         return $form;
